Refactoring
Moved all styled components outside of class components. 

Refactored privacy policy to map over an array of objects, stored separately.

Renamed all styling files.
