+++ src/components/auth/Login.js
@@ -1,18 +1,11 @@
 import React, { Component } from 'react';
 import { Typography, Row, Col } from "antd";
-import styled from "styled-components";
 
 import LoginForm from "./../forms/LoginForm";
-import { FullPage, Centered } from "./../layout/Layout";
+import { FullPage, Centered } from "./../layout/app_styles";
+import { BannerImage } from "./auth_styles";
 
 const { Title } = Typography;
-
-const BannerImage = styled.div`
-  background-image: url(https://foodladder.org/wp-content/uploads/2018/08/Frame-06-08-2018-02-23-43-2-1.jpg);
-  width: 100%;
-  height: 100%;
-  background-size: cover;
-`;
 
 export class Login extends Component {
   render() {
