--- src/components/auth/ProjectConfirmation.js
@@ -1,13 +1,0 @@
-import React, { Component } from "react";
-
-class ProjectConfirmation extends Component {
-
-  render() {
-
-    return(
-      <h1>Project Confirmation</h1>
-    );
-  }
-}
-
-export default ProjectConfirmation;
