--- src/components/auth/ProjectProfileDetails.js
@@ -1,13 +1,0 @@
-import React, { Component } from "react";
-
-class ProjectProfileDetails extends Component {
-
-  render() {
-
-    return(
-      <h1>Project Profile Details</h1>
-    );
-  }
-}
-
-export default ProjectProfileDetails;
