+++ src/components/auth/auth_styles.js
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const BannerImage = styled.div`
+  background-image: url(https://foodladder.org/wp-content/uploads/2018/08/Frame-06-08-2018-02-23-43-2-1.jpg);
+  width: 100%;
+  height: 100%;
+  background-size: cover;
+`;
