+++ src/components/forms/InvitationCodeForm.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import { reduxForm, Field } from "redux-form";
 import { Button, Card } from "antd";
-import { Input } from "./FormLayout";
+import { Input } from "./form_styles";
 
 class InvitationCodeForm extends Component {
