+++ src/components/forum/thread/post/post_styles.js
