+++ src/components/landing/Landing.js
@@ -1,19 +1,11 @@
 import React, { Component } from 'react';
-import styled from "styled-components";
 import { connect } from "react-redux";
 
-import { Centered, FullPage } from "./layout/Layout";
-import InvitationCodeForm from "./forms/InvitationCodeForm";
+import { Centered, FullPage } from "../layout/app_styles";
+import InvitationCodeForm from "../forms/InvitationCodeForm";
 
-import bannerImage from "./../assets/images/marketing_image_3.jpeg";
-
-import { setHeader, resetHeader } from "./../actions/index";
-
-const Banner = styled.div`
-  background-image: url(${bannerImage});
-  background-size: cover;
-  background-position: center;
-`;
+import { setHeader, resetHeader } from "../../actions/index";
+import { Banner } from "./landing_styles";
 
 export class Landing extends Component {
