+++ src/components/landing/landing_styles.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+import bannerImage from "./../../assets/images/marketing_image_3.jpeg";
+
+export const Banner = styled.div`
+  background-image: url(${bannerImage});
+  background-size: cover;
+  background-position: center;
+`;
