+++ src/components/misc/PrivacyPolicyStyles.js
@@ -1,0 +1,5 @@
+import styled from "styled-components";
+
+export const Group = styled.div`
+  padding-left: 24px;
+`;
